# Fix js & img tasks from logging being finished before actually being finished