--- a/lib/gulp/image.js
+++ b/lib/gulp/image.js
@@ -4,6 +4,7 @@ const newer = require('gulp-newer');
 const gulpSvgSprite = require('gulp-svg-sprite');
 const gulp = require('gulp');
 const plumber = require('gulp-plumber');
+const merge = require('merge-stream');
 const config = require('./config.js');
 
 // CSS task
@@ -16,7 +17,7 @@ function img() {
     return svgSprite({src: config.src_base_path + svg.src, dest: config.dest_base_path + svg.dest, name: svg.name});
   });
 
-  return Promise.all(imgTasks.concat(spriteTasks));
+  return merge(imgTasks.concat(spriteTasks));
 }
 
 // Optimize Images