--- a/lib/gulp/js.js
+++ b/lib/gulp/js.js
@@ -4,6 +4,7 @@ const presetEnv = require('@babel/preset-env');
 const concat = require('gulp-concat');
 const gulp = require('gulp');
 const plumber = require('gulp-plumber');
+const merge = require('merge-stream');
 const config = require('./config.js');
 
 function js() {
@@ -15,7 +16,7 @@ function js() {
     return jsConcat({src: config.src_base_path + js.src, dest: config.dest_base_path + js.dest, name: js.name});
   });
 
-  return Promise.all(jsTasks.concat(jsConcatTasks));
+  return merge(jsTasks.concat(jsConcatTasks));
 }
 
 function jsCompile({src, dest, browserSync = false}) {