This is actually also a fix for #22, but for the `js` and `img` tasks.

**Before**
- `js` & `img` duration times are incorrect
- `js` & `img` seemed to be finished before they are completed (`gulp-imagemin` outputs after `build` task)
```
[14:08:14] Working directory changed to ~/workspace/buildozer
[14:08:17] Using gulpfile ~/workspace/buildozer/gulpfile.js
[14:08:17] Starting 'build'...
[14:08:17] Starting 'clean'...
[14:08:17] Finished 'clean' after 7.52 ms
[14:08:17] Starting 'copy'...
          Files copied:
          - /Users/martijn.cuppens/workspace/buildozer-test/copy/file.txt to
            /Users/martijn.cuppens/workspace/buildozer-test/dest/file.txt
[14:08:17] Finished 'copy' after 61 ms
[14:08:17] Starting 'css'...
[14:08:17] Starting 'js'...
[14:08:17] Starting 'img'...
[14:08:17] Finished 'js' after 92 ms
[14:08:17] Finished 'img' after 92 ms
[14:08:17] gulp-imagemin: Minified 0 images
[14:08:17] Finished 'css' after 337 ms
[14:08:17] Finished 'build' after 410 ms
[14:08:21] gulp-imagemin: Minified 3 images (saved 200 kB - 14.4%)
✨  Done in 7.64s.
```

**After**
- Timings are correct
- Logging happens in the correct order
```
[14:37:35] Working directory changed to ~/workspace/buildozer
[14:37:36] Using gulpfile ~/workspace/buildozer/gulpfile.js
[14:37:36] Starting 'build'...
[14:37:36] Starting 'setProduction'...
[14:37:36] Finished 'setProduction' after 650 μs
[14:37:36] Starting 'clean'...
[14:37:36] Finished 'clean' after 6.38 ms
[14:37:36] Starting 'copy'...
          Files copied:
          - /Users/martijn.cuppens/workspace/buildozer-test/copy/file.txt to
            /Users/martijn.cuppens/workspace/buildozer-test/dest/file.txt
[14:37:36] Finished 'copy' after 27 ms
[14:37:36] Starting 'css'...
[14:37:36] Starting 'js'...
[14:37:36] Starting 'img'...
[14:37:36] gulp-imagemin: Minified 0 images
[14:37:36] Finished 'js' after 309 ms
[14:37:36] Finished 'css' after 316 ms
[14:37:40] gulp-imagemin: Minified 3 images (saved 200 kB - 14.4%)
[14:37:40] Finished 'img' after 4.1 s
[14:37:40] Finished 'build' after 4.14 s
✨  Done in 5.52s.
```